Remove fixed email and membership attribute detail
COmanage Registry might return multiple emails for
a COPerson. The email to select and use can now be
input to the constructor in the form of a regular
expression object. The default is a regular expression
object that selects either 'official' or 'delivery'.

The on the wire encoding of the membership attribute
can now be set in the constructor. The default is
urn:oid:1.3.6.1.4.1.5923.1.5.1.1, which has a
friendly name of 'isMemberOf' from the eduPerson
specification.

Also added logic to handle missing information to
prevent exceptions.